requirements.txt
streamlit==1.31.0
pydantic==2.5.3
langchain==0.2.11
langchain-openai==0.1.17
langchain-core==0.2.23
azure-search-documents==11.4.0
python-dotenv==1.0.0

# streamlit==1.31.0
# langchain==0.2.11
# langchain-openai==0.1.17
# langchain-core==0.2.23
# azure-search-documents==11.4.0
# azure-core==1.29.5
# azure-identity==1.15.0
# python-dotenv==1.0.0
# pydantic==2.5.3
# pydantic_core==2.14.6
# openai==1.35.7
# pandas==2.1.4
# numpy==1.26.3